refactor(utils): extract shared token utilities
