+++ .changeset/refactor-token-helpers.md
@@ -1,0 +1,5 @@
+---
+"@lapidist/design-lint": patch
+---
+
+refactor: consolidate token helpers into shared utils module with JSDoc coverage
