+++ src/utils/tokens/index.ts
@@ -6,3 +6,11 @@
 export { wrapTokenError } from './wrap-token-error.js';
 export { parseTokensForTheme } from './parse-tokens-for-theme.js';
 export { normalizeTokens } from './normalize-tokens.js';
+export { normalizePath, type NameTransform } from './path.js';
+export { matchToken, closestToken, type TokenPattern } from './pattern.js';
+export {
+  flattenDesignTokens,
+  getFlattenedTokens,
+  type FlattenOptions,
+} from './flatten.js';
+export { extractVarName } from './css.js';
